Refactor module exports for better organization
