+++ laura_core/examples/perft.rs
@@ -19,7 +19,7 @@
 
 use std::str::FromStr;
 
-use laura_core::{movegen::*, Board};
+use laura_core::*;
 
 /// A collection of standardized perft test positions for the Laura-Core move generation.
 /// These positions are used to validate move generation correctness and measure performance.
